move helper modules to actors folder
