+++ FYOUNET/actors/OVRInput.ts
@@ -2,7 +2,7 @@
 import { Address } from "../actorsystem/types.ts";
 import { actorManager } from "../actorsystem/actorManager.ts";
 import { ActorP2P } from "../actorsystem/actorP2P.ts";
-import { OVRInterface } from "../actorsystem/helpermodules/OVRInterface.ts";
+import { OVRInterface } from "./helpermodules/OVRInterface.ts";
 
 interface OverlayPayload {
     addr: Address<OverlayPayload>;
